+++ gae_embedding.py
@@ -89,7 +89,7 @@ def GAEembedding(z, adj, args):
     # featrues from z
     # Louvain
     features = z
-    features = torch.FloatTensor(features)
+    features = torch.DoubleTensor(features)
 
     # Old implementation
     # adj, features, y_test, tx, ty, test_maks, true_labels = load_data(args.dataset_str)
@@ -118,7 +118,7 @@ def GAEembedding(z, adj, args):
     adj_norm = preprocess_graph(adj)
     adj_label = adj_train + sp.eye(adj_train.shape[0])
     # adj_label = sparse_to_tuple(adj_label)
-    adj_label = torch.FloatTensor(adj_label.toarray())
+    adj_label = torch.DoubleTensor(adj_label.toarray())
 
     pos_weight = float(adj.shape[0] * adj.shape[0] - adj.sum()) / adj.sum()
     norm = adj.shape[0] * adj.shape[0] / float((adj.shape[0] * adj.shape[0] - adj.sum()) * 2)
